move billing functionality to a separate trait
